mvp1 float32

+++ src/envs/amod_env_multi.py
@@ -303,16 +303,16 @@ class AMoD:
                                 dist['wages'], 
                                 size=int(d_original), 
                                 p=dist['probabilities']
-                            )
+                            ).astype(np.float32, copy=False)
                         else:
                             # Fallback to uniform wage if region not in distribution
-                            passenger_wages = np.full(int(d_original), self.wage)
+                            passenger_wages = np.full(int(d_original), self.wage, dtype=np.float32)
                         
                         # Track sampled wages for system-wide average
                         self.system_wage_samples.extend(passenger_wages.tolist())
                         
                         # Vectorized income effect calculation: city_avg_wage / passenger_wage
-                        income_effects = self.city_avg_wage / passenger_wages
+                        income_effects = np.float32(self.city_avg_wage) / passenger_wages
                         
                         # Vectorized utility calculations for all passengers
                         U_0_batch = (
@@ -327,7 +327,7 @@ class AMoD:
                             - income_effects * self.choice_price_mult * pr1
                             + self.bm_gamma * self.brand_momentum[1]
                         )
-                        U_reject_batch = np.full(int(d_original), U_reject)
+                        U_reject_batch = np.full(int(d_original), U_reject, dtype=np.float32)
                         
                         # Vectorized probability calculation
                         exp_utilities_batch = np.column_stack([
@@ -353,7 +353,7 @@ class AMoD:
                         
                     else:
                         # Uniform wage path
-                        passenger_wages = np.full(int(d_original), self.wage)
+                        passenger_wages = np.full(int(d_original), self.wage, dtype=np.float32)
                         self.system_wage_samples.extend(passenger_wages.tolist())
 
                         # Compute utilities (same for all passengers under uniform wage)
